+++ pkg/web/webserver.go
@@ -26,6 +26,7 @@ func GetGinEngine() *gin.Engine {
 				return
 			}
 			c.Header("X-Frame-Options", "DENY")
+			c.Header("Access-Control-Allow-Origin", "*")
 			// c.Header("Content-Security-Policy", "default-src 'self'; connect-src *; font-src *; script-src-elem * 'unsafe-inline'; img-src * data:; style-src * 'unsafe-inline';")
 			c.Header("X-XSS-Protection", "1; mode=block")
 			c.Header("Strict-Transport-Security", "max-age=31536000; includeSubDomains; preload")
@@ -34,6 +35,10 @@ func GetGinEngine() *gin.Engine {
 			c.Header("Permissions-Policy", "geolocation=(),midi=(),sync-xhr=(),microphone=(),camera=(),magnetometer=(),gyroscope=(),fullscreen=(self),payment=()")
 			c.Next()
 		})
+		ginEngine.GET("/users", GetUserHandler)
+		ginEngine.GET("/user/:uid", GetUserByUIDHandler)
+		ginEngine.GET("/vehicle/:plate", GetSingleVehicleHandler)
+		ginEngine.GET("/vehicles", GetVehiclesHandler)
 
 		logger.Info("Gin web server initialization complete.")
 	})