+++ CHANGELOG.md
@@ -1,11 +1,12 @@
-[v0.1.5] 2024/10/28
+[v0.2.5] 2024/11/11
 
-Feature: AMQP Support; Lua Scripts Init; Analysis Module
+Feature: AMQP Support; Lua Scripts Init; Analysis Module; Admin Backned API
 
 - Added basic AMQP and AMQP exchange control
 - Added Analysis Module Framework
 - Added caller module of Persistent DB
 - Added Direct XML Report
+- Added Backend API for Vehicle/User listing
 
 Update: Adapt bining simulation.go to conform to new lua script loading method